refactor: change arg type to "uint128" in interfaces
docs: explain the new hooks in the cancel and withdraw functions' NatSpec
refactor: emit event at the bottom of "_cancel" and "_withdraw"
test: add helper "createDefaultDaiStreamWithRecipient"
test: add helper "createDefaultDaiStreamWithSender"
test: add new test contracts needed for hooks test
test: improve the wording in function names and comments
test: order users alphabetically
test: reder to "msg.sender" as "caller" in comments
test: test the sender and the recipient hooks functionality

+++ src/interfaces/ISablierV2Recipient.sol
@@ -17,8 +17,8 @@ interface ISablierV2Recipient {
     function onStreamCanceled(
         uint256 streamId,
         address caller,
-        uint256 withdrawAmount,
-        uint256 returnAmount
+        uint128 withdrawAmount,
+        uint128 returnAmount
     ) external;
 
     /// @notice Reacts to a withdrawal from a stream.
@@ -31,6 +31,6 @@ interface ISablierV2Recipient {
     function onStreamWithdrawn(
         uint256 streamId,
         address caller,
-        uint256 withdrawAmount
+        uint128 withdrawAmount
     ) external;
 }
